Shift configs into realtime pkg
