+++ zk/realtime/kafka/config.go
@@ -1,0 +1,9 @@
+package kafka
+
+type KafkaConfig struct {
+	BootstrapServers []string
+	BlockTopic       string
+	TxTopic          string
+	ErrorTopic       string
+	ClientID         string
+}
